seedsファイル拡充 productsレコードを100追加
seedデータの拡充。
productsレコードをさらに100追加し、取引済みと未取引で分けました(transaction_recordsとも連携)。
