+++ db/seeds.rb
@@ -34,52 +34,15 @@ end
 end
 
 # products
-100.times do |n|
-  Product.create(
-    user_id: Faker::Number.within(range: 1..2),
-    name: Faker::Name.product_name,
-    description: Faker::Name.description,
-    size_id: Faker::Number.within(range: 1..10),
-    status_id: 1,
-    condition_id: Faker::Number.within(range: 1..6),
-    delivery_expense_id: Faker::Number.within(range: 1..2),
-    delivery_method_id: Faker::Number.within(range: 1..9),
-    area_id: Faker::Number.within(range: 1..47),
-    shipdate_id: Faker::Number.within(range: 1..3),
-    price: Faker::Number.within(range: 300..60000),
-    brand_id: Faker::Number.within(range: 1..30),
-    category_id: Faker::Number.within(range: 72..300)
-  )
-end
+require './db/seeds/products.rb'
 
 # photos
-100.times do |n|
-  Photo.create(
-    product_id: Faker::Number.unique.within(range: 1..100),
-    photo: open("#{Rails.root}/db/fixtures/img#{Faker::Number.within(range: 1..2)}.png")
-  )
-  Photo.create(
-    product_id: Faker::Number.within(range: 1..100),
-    photo: open("#{Rails.root}/db/fixtures/img#{Faker::Number.within(range: 1..2)}.png")
-  )
-  Photo.create(
-    product_id: Faker::Number.within(range: 1..100),
-    photo: open("#{Rails.root}/db/fixtures/img#{Faker::Number.within(range: 1..2)}.png")
-  )
-  Photo.create(
-    product_id: Faker::Number.within(range: 1..100),
-    photo: open("#{Rails.root}/db/fixtures/img#{Faker::Number.within(range: 1..2)}.png")
-  )
-  Photo.create(
-    product_id: Faker::Number.within(range: 1..100),
-    photo: open("#{Rails.root}/db/fixtures/img#{Faker::Number.within(range: 1..2)}.png")
-  )
-end
+require './db/seeds/photos.rb'
 
 # transaction_records
-20.times do |n|
+100.times do |n|
   TransactionRecord.create(
     user_id: Faker::Number.within(range: 1..2),
-    product_id: Faker::Number.unique.within(range: 1..20)
+    product_id: n + 1
   )
 end
